backend/requirements.txt
fastapi[standard]>=0.115.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
redis>=4.5.4
celery>=5.2.7
pyjwt>=2.8.0
pwdlib[argon2]
authlib>=1.6.6
itsdangerous
python-multipart
boto3
uvicorn[standard]
flower
pydantic-settings
httpx
geoip2
cryptography
psycopg2-binary
torch>=2.0.0
torchvision>=0.19.0
onnxruntime>=1.20.0
opencv-python-headless>=4.10.0
Pillow>=11.0
numpy>=2.0
scikit-image>=0.24.0
diffusers>=0.30.0
transformers>=4.40.0
accelerate>=0.30.0
